Move playbook mocks back into playbook tests
