+++ playbook/playbook.go
@@ -3,10 +3,11 @@ package playbook
 import (
 	"errors"
 	"fmt"
-	"gopkg.in/yaml.v2"
 	"io/ioutil"
 	"os"
 	"path/filepath"
+
+	"gopkg.in/yaml.v2"
 )
 
 type Meta struct {
